mysite/requirements.txt
django==2.2
pytest-django==4.5.2
djangorestframework==3.13.1
Markdown==3.3.7
django-filter==21.1